refactor: MemberE2ETest 리팩터
